add pipelines blueprint
