+++ app/pipelines/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+pipelines_bp = Blueprint('pipelines', __name__)
+
+from app.pipelines import routes
